requirements.txt
Flask==1.0.2
flask-restplus==0.12.1
google-cloud-datastore